third_party/WebKit/LayoutTests/crypto/sign-verify-expected.txt

Tests cypto.subtle.sign and crypto.subtle.verify

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.verify(hmacSha1, keys.hmacSha1, null, data) threw exception TypeError: Type error.
PASS crypto.subtle.verify(hmacSha1, keys.hmacSha1, 'a', data) threw exception TypeError: Type error.
PASS crypto.subtle.verify(hmacSha1, keys.hmacSha1, [], data) threw exception TypeError: Type error.
PASS crypto.subtle.sign({name: 'sha-1'}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: {name: 'sha-256'}}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac'}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: 3}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: null}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: {}}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: {name: 'foo'}}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'hmac', hash: {name: 'AES-CBC'}}, keys.hmacSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5'}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: 3}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: null}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {}}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'foo'}}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.sign({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'AES-CBC'}}, keys.rsaSsaSha1, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS signResult.byteLength is 17
PASS verifyResult is true
PASS verifyResult is false
PASS successfullyParsed is true

TEST COMPLETE
